Working with ticks and asynchronous methods
I think I need to review async/await further. Calling these within a synchronous method makes some sense, everything is queued for the beginning of the next event loop cycle except for `setTimeout` which gets queued onto the end of the following cycle.

Calling these within an async method appears to make the program wait for each method to execute before proceeding, except for the `setTimeout` method, unless it's contained within an awaited promise and the resolution is in the timeout...